==> components/risk-engine/notebooks/00_fetch_data.ipynb <==
# README
This notebook fetches all data required for the risk engine and exports 3 dataframes:
- risk_cust_attributes
- risk_cust_fixed_lists
- risk_cust_trx
  
Note: This notebook will not be turned to a production script (at least inside the risk engine). This shall be replaced by the API call itself either for:
- scoring or rescoring.
- predict mode vs train mode
  
This notebook shall help the BE team to develop the respective DBs and API endpoints.

# 1- Imports

import numpy as np
import pandas as pd
import yaml
import sys
import os
from pathlib import Path
import re

# paths and custom modules
cwd_path = Path(os.getcwd())
data_path = cwd_path / 'data'
sys.path.append(str(cwd_path))
from utils import return_catalog, dataFetchingFunc, strip_strings, stnd_missing_values_type, build_spark_session
data_catalog, models_catalog = return_catalog()

### Export version

data_date = '20240810'
analysis_date = '20240810'

# 2- Fetch

### Filtered train data

query = '''
WITH 
ordered_orders AS (
    SELECT 
        fo.ClientID,
        fo.SSN, 
        fo.ContractDateTime,
        fo.OrderCode,
        ROW_NUMBER() OVER(PARTITION BY fo.ClientID, fo.SSN ORDER BY fo.ContractDateTime ASC) AS row_num
    FROM DWH_MiniCash.dbo.Fact_Orders fo 
    WHERE OrderStatusID IN (8, 12)
),

filtered_orders AS (
    SELECT 
        ClientID,
        SSN,
        ContractDateTime,
        OrderCode
    FROM ordered_orders
    WHERE row_num = 1
),

joined_orders AS (
    SELECT 
        sfo.ClientID, 
        sfo.SSN,
        sfo.ContractDateTime, 
        sfo.OrderCode, 
        fia.RequerdDate, 
        fia.PaymentDate,
        CASE 
            WHEN PaymentDate IS NOT NULL THEN DATEDIFF(DAY, fia.RequerdDate, fia.PaymentDate) 
            ELSE DATEDIFF(DAY, fia.RequerdDate, GETDATE()) 
        END AS DaysDifference,
        CASE 
            WHEN DATEDIFF(DAY, fia.RequerdDate, ISNULL(fia.PaymentDate, GETDATE())) > 90 THEN 1
            ELSE 0
        END AS IsAbove90Days,
        fia.Amount,
        fia.BenefitValue
    FROM filtered_orders sfo
    LEFT JOIN DWH_MiniCash.dbo.Fact_InstallmentAmount fia
    ON sfo.OrderCode = fia.OrderCode
),

grouped_orders AS (
    SELECT 
        jo.ClientID, 
        jo.SSN,
        jo.OrderCode, 
        jo.ContractDateTime, 
        SUM(jo.IsAbove90Days) AS def_flag,
        o.MaxNumberOfPayment AS 'tenor',
        o.OrderID,
        SUM(jo.Amount) AS order_amount
    FROM joined_orders jo
    LEFT JOIN DWH_MiniCash.dbo.Fact_Orders o 
    ON o.ordercode = jo.OrderCode
    GROUP BY jo.ClientID, jo.SSN, jo.ContractDateTime, jo.OrderCode, o.MaxNumberOfPayment, o.OrderID
),

lvl6 AS (
    SELECT 
        SUM(Premium) 'price',
        OrderID 
    FROM DWH_MiniCash.dbo.Fact_OrderDetails 
    WHERE OrderStatusID IN (8,12) 
    GROUP BY OrderID
),

lvl7 AS (
    SELECT 
        SUM(BenefitValue) 'Benfit',
        OrderID 
    FROM DWH_MiniCash.dbo.Fact_OrderDetails 
    WHERE OrderStatusID IN (8,12) 
    GROUP BY OrderID
),

lvl8 AS (
    SELECT 
        SUM(PaymentAmount) 'Total_price',
        OrderID 
    FROM DWH_MiniCash.dbo.Fact_OrderDetails 
    WHERE OrderStatusID IN (8,12) 
    GROUP BY OrderID
)

SELECT      
    go.ClientID, 
    go.SSN, 
    go.ContractDateTime,
    -- ISNULL(go.order_amount,0) AS 'order_amount_t1' ,
    ROUND((SELECT lvl8.Total_price FROM lvl8 WHERE lvl8.orderid = go.orderid) * go.tenor,0) 'order_amount',
    ROUND((SELECT lvl7.Benfit FROM lvl7 WHERE lvl7.orderid = go.orderid) * go.tenor, 0) AS order_benefit,
    go.tenor,
    CASE
        WHEN go.def_flag > 0 THEN 1
        ELSE 0
    END AS def_flag 
FROM grouped_orders go
'''

first_loan_df = dataFetchingFunc(
    Query=query,
    SERVER="BI-DR-DB",
    DATABASE="master"
)

first_loan_df.head()

### Income data

query = '''
WITH income_v AS (
    SELECT 
        ClientID, 
        CreateDate,
        ABS(ClientSalary) + ABS(ClientOtherSalary) AS net_income,
        ABS(TotalOtherBurden) + ABS(TotalFamilyBurden) AS net_burden,
        ROW_NUMBER() OVER(PARTITION BY ClientID ORDER BY CreateDate) AS rn,
        ROW_NUMBER() OVER(PARTITION BY ClientID ORDER BY CreateDate DESC) AS rn_desc
    FROM 
        DWH_MiniCash.dbo.Dim_ClientCreditLimit ccl
    WHERE 
        ABS(ClientSalary) + ABS(ClientOtherSalary) > 100 AND
        ABS(ClientSalary) + ABS(ClientOtherSalary) <= 150000
        AND CreateDate >= '2019-01-01'
)
SELECT 
    ClientID,
    MIN(CreateDate) AS first_update_date,
    MAX(CASE WHEN rn = 1 THEN net_income END) AS net_income_first,
    MAX(CASE WHEN rn = 1 THEN net_burden END) AS net_burden_first,
    MAX(CreateDate) AS last_update_date,
    MAX(CASE WHEN rn_desc = 1 THEN net_income END) AS net_income_last,
    MAX(CASE WHEN rn_desc = 1 THEN net_burden END) AS net_burden_last
FROM 
    income_v
GROUP BY 
    ClientID
'''

income_burden_df = dataFetchingFunc(
    Query=query,
    SERVER="BI-DR-DB",
    DATABASE="master"
)

income_burden_df

### Application data

query = '''
SELECT oad.ClientNationalNumber AS SSN, ht.HouseTypeName, oad.SportingClubId,
	oad.CarTypeId, oad.CarModel, oad.ChilderenCount 
FROM Lnd_MiniCash.dbo.OrderAdditionalData oad
	INNER JOIN (SELECT ClientNationalNumber, MAX(Id) AS max_id
	FROM Lnd_MiniCash.dbo.OrderAdditionalData oad
	GROUP BY ClientNationalNumber) sub ON sub.max_id = oad.Id 
LEFT JOIN Lnd_MiniCash.dbo.HouseType ht ON oad.HousingTypeId = ht.HouseTypeID
'''

app_d1_df = dataFetchingFunc(
    Query=query,
    SERVER="BI-DR-DB",
    DATABASE="master"
)

app_d1_df

query = """
SELECT dc.ClientID, dc.SSN, dc.MobilePhone AS 'phone_number_1',
dc.HomePhone AS 'phone_number_2', dc.JobName, ms.MartialStatusName, dc.ClientClass AS repayment_class
FROM DWH_MINICASH.DBO.DIM_CLIENTS DC
LEFT JOIN LND_MINICASH.DBO.MARTIALSTATUS MS ON MS.MARTIALSTATUSID = DC.MARTIALSTATUSID
ORDER BY DC.CLIENTID DESC
"""

app_d2_df = dataFetchingFunc(
    Query=query,
    SERVER="BI-DR-DB",
    DATABASE="master"
)

app_d2_df

query = """
SELECT j.jobsName AS job_name, j.EnName AS job_name_english, jt.Name AS job_type, j.RiskId AS job_risk_id
FROM Btech_live.dbo.Jobs j
LEFT JOIN Btech_live.dbo.JobType jt ON j.JobTypeId = jt.Id
"""

app_d3_df = dataFetchingFunc(
    Query=query,
    SERVER="10.3.1.240",
    USERNAME="sa",
    PASSWORD="P@ssw0rd",
    DATABASE="Btech_Live"
)

app_d3_df

### Address data

# query = """
# WITH max_orders AS (
# SELECT ClientID, MAX(OrderDate) AS max_order_date
# FROM Btech_live.dbo.Orders o 
# GROUP BY ClientID
# )
# SELECT o.ClientID, c.SSN, 
# -- o.OrderID, 
# g.GovernorateAr AS address_governrate, 
# d.DistrictNameAr AS address_city,
# a.AreaNameAr AS address_area
# FROM Btech_live.dbo.Orders o
# INNER JOIN max_orders on (max_orders.ClientID = o.ClientID) AND (max_orders.max_order_date = o.OrderDate)
# LEFT JOIN Btech_live.dbo.Client c ON c.ClientID = o.ClientID 
# LEFT JOIN  Btech_live.dbo.Area a WITH (nolock) on a.AreaID = o.DistrictAreaId
# LEFT JOIN Btech_live.dbo.District d on d.DistrictID = a.DistrictID
# LEFT JOIN Btech_live.dbo.Governorate g on g.GovernID = d.GovernorateID
# LEFT JOIN Btech_live.dbo.OrderFulfillment of2 ON o.OrderID = of2.MCOrderID
# ORDER BY o.ClientID DESC
# """

# cust_address_df = dataFetchingFunc(
#     Query=query,
#     SERVER="10.3.1.240",
#     USERNAME="sa",
#     PASSWORD="P@ssw0rd",
#     DATABASE="Btech_Live"
# )

# cust_address_df

query = """
WITH max_orders AS
  (SELECT ClientID,
          MAX(OrderDate) AS max_order_date
   FROM Btech_live.dbo.Orders o
   GROUP BY ClientID)
SELECT o.ClientID,
       c.SSN,
       -- o.OrderID,
       g.GovernID AS GovernorateID,
       d.DistrictID AS CityID,
       a.AreaID AS AreaID
FROM Btech_live.dbo.Orders o
INNER JOIN max_orders ON (max_orders.ClientID = o.ClientID)
AND (max_orders.max_order_date = o.OrderDate)
LEFT JOIN Btech_live.dbo.Client c ON c.ClientID = o.ClientID
LEFT JOIN Btech_live.dbo.Area a WITH (nolock) ON a.AreaID = o.DistrictAreaId
LEFT JOIN Btech_live.dbo.District d ON d.DistrictID = a.DistrictID
LEFT JOIN Btech_live.dbo.Governorate g ON g.GovernID = d.GovernorateID
LEFT JOIN Btech_live.dbo.OrderFulfillment of2 ON o.OrderID = of2.MCOrderID
ORDER BY o.ClientID DESC
"""

cust_address_df = dataFetchingFunc(
    Query=query,
    SERVER="10.3.1.240",
    USERNAME="sa",
    PASSWORD="P@ssw0rd",
    DATABASE="Btech_Live"
)

addresss_map = pd.read_excel(data_path+f'feed/live_lists/20240828_mylo_addresses.xlsx', sheet_name='main')

cust_address_df = cust_address_df.merge(
        addresss_map,
        left_on= ['AreaID', 'CityID','GovernorateID'],
        right_on=['mc_AreaID', 'mc_CityID', 'mc_GovernorateID'],
        how='left',
        indicator=True
    )

cust_address_df = cust_address_df[['ClientID', 'SSN', 'GovernorateEn', 'CityNameEn', 'AreaNameEn']]
cust_address_df.dropna(subset=['GovernorateEn', 'CityNameEn', 'AreaNameEn'], inplace=True)
cust_address_df.rename(columns={'GovernorateEn':'address_governorate',
                        'CityNameEn':'address_city',
                        'AreaNameEn':'address_area'}, inplace=True)

### I-Score data

query = '''
SELECT
    SSN,
    Score AS i_score,
    IscoreReport
FROM Lnd_MiniCash.dbo.IscoreReport
'''

iscore_df = dataFetchingFunc(
    Query=query,
    SERVER="BI-DR-DB",
    DATABASE="master"
)

iscore_df

### Live lists data

# for b2b customers
b2b_customers_df = pd.read_excel(data_path+'feed/live_lists/20240527_b2b_customer_list.xlsx', dtype=str)[['customer_phone']]
b2b_customers_df = strip_strings(b2b_customers_df)
b2b_cust_list = ', '.join(f"'{phone}'" for phone in b2b_customers_df['customer_phone'].tolist())
b2b_customers_df.head()

# for salary scales (used only to correct current salaries)
salary_scales_df = pd.read_excel(data_path+'feed/live_lists/20240401_salary_scales.xlsx')
salary_scales_df.head()

# for mc to mylo job-mapping
mylo_jobs_df = pd.read_excel(data_path+'feed/live_lists/20240620_job_mapping.xlsx')
mylo_jobs_df.head()

### Transactions

query = f"""
SELECT [Customer Phone] AS customer_phone,
       [SSoldDateID],
       [Online or  Branch] AS online_vs_branch,
       [Transaction type] AS trx_type,
       [Net Sales] AS net_sales,
       [Quantity] AS net_qty
FROM [DWH_Sales].[dbo].[Fact_online_Sales]
WHERE [Customer Type] = 'StandardCP'
  AND Business_LineValue IN ('B Tech X',
                              'BTECH Mini',
                              'C_CHAINS_MDA',
                              'C_CHAINS_SDA',
                              'C_LED',
                              'C_MDA',
                              'C_SDA',
                              'R_CallCent',
                              'R_Online',
                              'R_Outlet',
                              'R_Stores')
  AND [Customer Phone] NOT IN ({b2b_cust_list})					  
"""

trx_df = dataFetchingFunc(
    Query=query,
    SERVER="BI2-DR-DB",
    DATABASE="master"
)

trx_df

# 3- Merge & Formatting

# Strip all strings
first_loan_df = strip_strings(first_loan_df)
income_burden_df = strip_strings(income_burden_df)
salary_scales_df = strip_strings(salary_scales_df)
mylo_jobs_df = strip_strings(mylo_jobs_df)
iscore_df = strip_strings(iscore_df)
app_d1_df = strip_strings(app_d1_df)
app_d2_df = strip_strings(app_d2_df)
app_d3_df = strip_strings(app_d3_df)
cust_address_df = strip_strings(cust_address_df)
trx_df = strip_strings(trx_df)

# Merge customer's attributes
cust_attr_df = first_loan_df.merge(income_burden_df, how='left', on='ClientID', indicator='indc_income_burden_df')\
    .merge(iscore_df, how='left', on='SSN', indicator='indc_iscore_df')\
        .merge(app_d1_df, how='left', on='SSN', indicator='indc_app_d1_df')\
            .merge(app_d2_df, how='left', on='ClientID', indicator='indc_app_d2_df')\
                .merge(app_d3_df, how='left', left_on='JobName', right_on='job_name', indicator='indc_app_d3_df')\
                    .merge(salary_scales_df, how='left', left_on='JobName', right_on='ss_job_name', indicator='indc_salary_scales')\
                        .merge(cust_address_df, on='ClientID', how='left', indicator='indc_cust_address_df')\
                            .merge(mylo_jobs_df, on='job_name', how='left', indicator='indc_mylo_jobs_df')
   
 
cust_attr_df.head()

# def-sequential cleaning (trx)
def base_clean_for_train_trx(trx_df):
    
    trx_df = trx_df.rename(columns={'SSoldDateID': 'trx_date'})
    trx_df = trx_df.rename(columns={'trx_type': 'sales_vs_returns'})
    
    trx_df = trx_df.astype({'net_sales': 'float64'})
    trx_df = trx_df.astype({'net_qty': 'float64'})
    trx_df['trx_date'] = pd.to_datetime(trx_df['trx_date'], format='%Y%m%d')
    
    return trx_df

trx_df_l2 = base_clean_for_train_trx(trx_df.copy())
trx_df_l2.head()

# def-sequential cleaning (attributes)

def base_clean_for_train_attr(df):

    df = df.rename(columns={'ClientID': 'client_id'})
    df = df.drop_duplicates(subset=['client_id'])
    df = df.rename(columns={'SSN_x': 'ssn'})
    df = df.rename(columns={'ContractDateTime': 'contract_date'})
    df = df.rename(columns={'tenor': 'first_ord_tenor'})
    df = df.dropna(subset=['order_amount'])
    df = df.rename(columns={'order_amount': 'first_ord_amount'})
    df = df.rename(columns={'order_benefit': 'first_ord_benefit'})
    df = df.dropna(subset=['contract_date'])
    df = df.astype({'first_ord_amount': 'float64'})
    df = df.astype({'first_ord_benefit': 'float64'})
    df = df.dropna(subset=['first_update_date'])
    df = df.astype({'first_update_date': 'datetime64[ns]'})
    df = df.rename(columns={'first_update_date': 'first_income_up_date'})
    df = df.rename(columns={'last_update_date': 'last_income_up_date'})
    df = df.astype({'last_income_up_date': 'datetime64[ns]'})
    df = df.drop(columns=['indc_income_burden_df'])
    df = df.rename(columns={'def_flag': 'fo_par90_flag'})
    df = df.rename(columns={'i_score': 'iscore_score'})
    df = df.rename(columns={'IscoreReport': 'iscore_report'})
    df = df.rename(columns={'is_iScoreNohitConsumer': 'is_iscore_nohit'})
    df = df.rename(columns={'HouseTypeName': 'house_type'})
    df = df.rename(columns={'SportingClubId': 'club_level'})
    df = df[df['indc_app_d1_df'].astype("string") == "both"]
    df = df.rename(columns={'CarTypeId': 'car_type_id'})
    df = df.rename(columns={'CarModel': 'car_model_year'})
    df = df.rename(columns={'ChilderenCount': 'children_count'})
    df = df.drop(columns=['indc_app_d1_df', 'SSN_y'])
    df = df.drop(columns=['JobName'])
    df = df.rename(columns={'MartialStatusName': 'marital_status'})
    df = df.drop(columns=['indc_app_d2_df'])
    df = df.drop(columns=['indc_app_d3_df'])
    df = df.drop(columns=['indc_salary_scales'])
    
    df['contract_date'] = pd.to_datetime(df['contract_date']).dt.date 
    df = df.astype({'contract_date': 'datetime64[ns]'})
    df['last_income_up_date'] = pd.to_datetime(df['last_income_up_date']).dt.date
    df = df.astype({'last_income_up_date': 'datetime64[ns]'})
    df['club_level'] = df['club_level'].astype('float64').astype('Int64')
    df['car_type_id'] = df['car_type_id'].astype('float64').astype('Int64')
    df['car_model_year'] = df['car_model_year'].astype('float64').astype('Int64')
    df['children_count'] = df['children_count'].astype('float64').astype('Int64')
    df = df.rename(columns={'repayment_class': 'current_repayment_class'})
    df = df.drop(columns=['ss_job_name'])
    
    return df

cust_attr_df_l2 = base_clean_for_train_attr(cust_attr_df.copy())
cust_attr_df_l2.head()

# view after def-sequential cleaning
print(cust_attr_df_l2.shape)
print(cust_attr_df_l2['fo_par90_flag'].value_counts(normalize=True))

# 4- Flags and other attributes

# std copy
cust_attr_df_lf = cust_attr_df_l2.copy()

# adding flags
cust_attr_df_lf['flag_is_mc_customer'] = 1
cust_attr_df_lf['flag_is_prv_cash_trx'] = 0
cust_attr_df_lf['flag_is_rescore'] = 0

# adding dummy address atributes
# cust_attr_df_lf['address_governorate'] = 'default_value'
# cust_attr_df_lf['address_city'] = 'default_value'
# cust_attr_df_lf['address_area'] = 'default_value'

# adding other attributes
cust_attr_df_lf['insurance_type'] = 'default_value'
cust_attr_df_lf['mobile_os_type'] = 'default_value'

cust_attr_df_lf['net_income'] = cust_attr_df_lf['net_income_last']
cust_attr_df_lf['net_burden'] = cust_attr_df_lf['net_burden_last']

# re-ordering columns for readability
cust_attr_df_lf = cust_attr_df_lf[['client_id', 'ssn', 'phone_number_1', 'phone_number_2','flag_is_mc_customer', 
'flag_is_prv_cash_trx', 'flag_is_rescore', 'contract_date',
#'job_name',
'job_name_map', 
'job_type','net_income', 'net_burden','first_income_up_date','net_income_first', 'net_burden_first', 'last_income_up_date',
'net_income_last', 'net_burden_last',
'ss_mc_program', 'ss_min_income', 'ss_max_income', 'ss_initial_cl',
'job_map_min_salary',
'job_map_max_salary',
'insurance_type', 'marital_status', 'children_count', 'address_governorate', 'address_city', 'address_area',
'house_type', 'car_type_id', 'car_model_year', 'club_level', 'mobile_os_type', 'indc_iscore_df',
'iscore_score', 'iscore_report', 'current_repayment_class', 'first_ord_amount',  'first_ord_benefit','first_ord_tenor','fo_par90_flag']].copy()

# 5- EOD Formatting

def clean_eod_data_att(df):
    df = df.astype({'flag_is_mc_customer': 'int8'})
    df = df.astype({'flag_is_prv_cash_trx': 'int8', 'flag_is_rescore': 'int8'})
    df = df.astype({'net_income_first': 'float32', 'net_burden_first': 'float32', 'net_income_last': 'float32', 'net_burden_last': 'float32',  'net_income': 'float32', 'net_burden': 'float32'})
    df = df.astype({'ss_min_income': 'float32', 'ss_max_income': 'float32', 'ss_initial_cl': 'float32'})
    df = df.astype({'first_ord_amount': 'float32', 'first_ord_benefit': 'float32'})
    df = df.astype({'children_count': 'int8'})
    df = df.astype({'car_type_id': 'string'})
    df = df.astype({'iscore_score': 'float32'})
    df = df.astype({'first_ord_amount': 'float32'})
    df = df.astype({'first_ord_tenor': 'int16'})
    df = df.astype({'fo_par90_flag': 'int8'})
    df = df.astype({'phone_number_1': 'string', 'phone_number_2': 'string', 'club_level': 'string'})
    df = df.astype({'indc_iscore_df': 'string'})
    df['indc_iscore_df'] = df['indc_iscore_df'].str.replace("both", "1", case=False, regex=False)
    df['indc_iscore_df'] = df['indc_iscore_df'].str.replace("left_only", "0", case=False, regex=False)
    df = df.rename(columns={'indc_iscore_df': 'is_iscore'})
    df = df.astype({'is_iscore': 'int8'})
    
    df.loc[df['car_model_year'] == 0, 'car_model_year'] = np.NaN
    df['car_model_year'] = pd.to_datetime(df['car_model_year'], format='%Y', errors='coerce')
    df['car_model_year'] = pd.to_datetime(df['car_model_year'], format='%Y%m%d')
    df['job_map_min_salary'] = df['job_map_min_salary'].round(0)
    df['job_map_max_salary'] = df['job_map_max_salary'].round(0)
    return df

cust_attr_df_eod = clean_eod_data_att(cust_attr_df_lf.copy())
cust_attr_df_eod.head()

def clean_eod_data_trx(df):
    df = df.astype({'online_vs_branch': 'int8'})
    df = df.astype({'sales_vs_returns': 'int8'})
    df = df.astype({'net_sales': 'float32'})
    df = df.astype({'net_qty': 'float32'})
    return df

cust_trx_df_eod = clean_eod_data_trx(trx_df_l2.copy())
cust_trx_df_eod.head()

def convert_int8_to_int16(df):
    for col in df.columns:
        if df[col].dtype == 'Int8':
            df[col] = df[col].astype('Int16')
    return df

cust_attr_df_eod = convert_int8_to_int16(cust_attr_df_eod.copy())
cust_trx_df_eod = convert_int8_to_int16(cust_trx_df_eod.copy())

# Update Marital Status to English
mapping = {
    'متزوج': 'Married',
    'غير متزوج': 'Single',
    'مطلق': 'Divorced',
    'ارمل':'Widowed'
}
cust_attr_df_eod['marital_status'] = cust_attr_df_eod['marital_status'].replace(mapping)

# Update House Type to English
mapping = {
    'ايجار جديد': 'New Rent',
    'تمليك': 'Owned',
    'ايجار قديم': 'Old Rent'
}
cust_attr_df_eod['house_type'] = cust_attr_df_eod['house_type'].replace(mapping)

# standardize missing values
cust_attr_df_eod = stnd_missing_values_type(cust_attr_df_eod.copy())
cust_trx_df_eod = stnd_missing_values_type(cust_trx_df_eod.copy())

# 6- Export

--------------

cust_attr_df_eod.dtypes

cust_trx_df_eod.dtypes

# Save to_model customer's attributes
cust_attr_df_eod.to_parquet(data_path/f'feed/{data_date}_risk_cust_attributes.parquet')
# Save to_model customer's trx
# cust_trx_df_eod.to_parquet(data_path+f'feed/{data_date}_risk_cust_trx.parquet')

# Analytics/fetch

# query = '''
# WITH RankedInquiries AS (
#     SELECT 
#         ClientID,
#         SSN,
#         [Last Inquiry Date],
#         Class,
#         ROUND(CreditLimit, 0) AS CreditLimit,
#         ROUND(MaxCredit, 0) AS MaxCredit,
#         ROUND(Amount, 0) AS Amount,
#         ROW_NUMBER() OVER (PARTITION BY SSN ORDER BY [Last Inquiry Date] DESC) AS rn
#     FROM 
#         Lnd_MiniCash.dbo.ClinetClass2020
# ),
# RankedAmounts AS (
#     SELECT
#         ClientID,
#         ROUND(MaxAmount, 0) AS MaxAmount,
#         CreateDate,
#         ROW_NUMBER() OVER (PARTITION BY ClientID ORDER BY CreateDate DESC) AS rn
#     FROM
#         Lnd_MiniCash.dbo.ClientCreditLimit
# ),
# FinalResults AS (
#     SELECT 
#         ri.SSN AS ssn,
#         ri.ClientID AS client_id,
#         ri.[Last Inquiry Date] AS inquiry_date,
#         ri.Class AS mc_repayment_class,
#         ri.MaxCredit AS mc_monthly_cl,
#         CASE
#             WHEN Class = 'A' THEN 1.5
#             WHEN Class IN ('B', 'B1') THEN 1.25
#             WHEN Class IN ('N', 'N1', 'N3', 'N4') THEN 1
#             WHEN Class IN ('C', 'C1', 'N2') THEN 0
#             ELSE NULL -- Optional: in case of unexpected class values
#         END AS mc_cl_rate,
#         c.mylo_migrated AS is_mylo_migrated,
#         ra.max_monthly_cl_pre_last,
#         ra.max_monthly_cl_last
#     FROM 
#         RankedInquiries AS ri
#     LEFT JOIN 
#         Lnd_MiniCash.dbo.Client c ON c.ClientID = ri.ClientID
#     LEFT JOIN (
#         SELECT
#             ClientID,
#             MAX(CASE WHEN rn = 2 THEN MaxAmount END) AS max_monthly_cl_pre_last,
#             MAX(CASE WHEN rn = 1 THEN MaxAmount END) AS max_monthly_cl_last
#         FROM
#             RankedAmounts
#         GROUP BY
#             ClientID
#     ) AS ra ON ra.ClientID = ri.ClientID
#     WHERE 
#         ri.rn = 1
# )
# SELECT 
#     ssn,
#     client_id,
#     inquiry_date,
#     mc_repayment_class,
#     ROUND(mc_monthly_cl, 0) AS mc_monthly_cl,
#     mc_cl_rate,
#     is_mylo_migrated,
#     ROUND(
#         CASE 
#             WHEN ROUND(mc_monthly_cl, 0) >= ROUND(max_monthly_cl_pre_last, 0) AND ROUND(mc_monthly_cl, 0) >= ROUND(max_monthly_cl_last, 0) THEN ROUND(mc_monthly_cl, 0)
#             WHEN ROUND(max_monthly_cl_pre_last, 0) >= ROUND(mc_monthly_cl, 0) AND ROUND(max_monthly_cl_pre_last, 0) >= ROUND(max_monthly_cl_last, 0) THEN ROUND(max_monthly_cl_pre_last, 0)
#             ELSE ROUND(max_monthly_cl_last, 0)
#         END, 0
#     ) AS mc_cl_migrated,
#     ROUND(
#         CASE 
#             WHEN CASE 
#                     WHEN ROUND(mc_monthly_cl, 0) >= ROUND(max_monthly_cl_pre_last, 0) AND ROUND(mc_monthly_cl, 0) >= ROUND(max_monthly_cl_last, 0) THEN ROUND(mc_monthly_cl, 0)
#                     WHEN ROUND(max_monthly_cl_pre_last, 0) >= ROUND(mc_monthly_cl, 0) AND ROUND(max_monthly_cl_pre_last, 0) >= ROUND(max_monthly_cl_last, 0) THEN ROUND(max_monthly_cl_pre_last, 0)
#                     ELSE ROUND(max_monthly_cl_last, 0)
#                 END IN (0, 1) THEN NULL
#             ELSE CASE 
#                     WHEN ROUND(mc_monthly_cl, 0) >= ROUND(max_monthly_cl_pre_last, 0) AND ROUND(mc_monthly_cl, 0) >= ROUND(max_monthly_cl_last, 0) THEN ROUND(mc_monthly_cl, 0)
#                     WHEN ROUND(max_monthly_cl_pre_last, 0) >= ROUND(mc_monthly_cl, 0) AND ROUND(max_monthly_cl_pre_last, 0) >= ROUND(max_monthly_cl_last, 0) THEN ROUND(max_monthly_cl_pre_last, 0)
#                     ELSE ROUND(max_monthly_cl_last, 0)
#                 END * mc_cl_rate * 18
#         END, 0
#     ) AS mylo_rev18_cl
# FROM 
#     FinalResults;
# '''

# cl_migration_df = dataFetchingFunc(
#     Query=query,
#     SERVER="BI-DR-DB",
#     DATABASE="master"
# )

# cl_migration_df.head()

query = '''
WITH RankedInquiries AS (
    SELECT 
        ClientID,
        SSN,
        [Last Inquiry Date],
        Class,
        CreditLimit,
        MaxCredit,
        Amount,
        ROW_NUMBER() OVER (PARTITION BY SSN ORDER BY [Last Inquiry Date] DESC) AS rn
    FROM 
        Lnd_MiniCash.dbo.ClinetClass2020
),
RankedAmounts AS (
    SELECT
        ClientID,
        MaxAmount,
        CreateDate,
        ROW_NUMBER() OVER (PARTITION BY ClientID ORDER BY CreateDate DESC) AS rn
    FROM
        Lnd_MiniCash.dbo.ClientCreditLimit
),
FinalResults AS (
    SELECT 
        ri.SSN AS ssn,
        ri.ClientID AS client_id,
        ri.[Last Inquiry Date] AS inquiry_date,
        ri.Class AS mc_repayment_class,
        ri.MaxCredit AS mc_monthly_cl,
        CASE
            WHEN Class = 'A' THEN 1.5
            WHEN Class IN ('B', 'B1') THEN 1.25
            WHEN Class IN ('N', 'N1', 'N3', 'N4') THEN 1
            WHEN Class IN ('C', 'C1', 'N2') THEN 0
            ELSE NULL
        END AS mc_cl_rate,
        c.mylo_migrated AS is_mylo_migrated,
        ra.max_monthly_cl_pre_last,
        ra.max_monthly_cl_last
    FROM 
        RankedInquiries AS ri
    LEFT JOIN 
        Lnd_MiniCash.dbo.Client c ON c.ClientID = ri.ClientID
    LEFT JOIN (
        SELECT
            ClientID,
            MAX(CASE WHEN rn = 2 THEN MaxAmount END) AS max_monthly_cl_pre_last,
            MAX(CASE WHEN rn = 1 THEN MaxAmount END) AS max_monthly_cl_last
        FROM
            RankedAmounts
        GROUP BY
            ClientID
    ) AS ra ON ra.ClientID = ri.ClientID
    WHERE 
        ri.rn = 1
)
SELECT 
    ssn,
    client_id,
    inquiry_date,
    mc_repayment_class,
    mc_cl_rate,
    is_mylo_migrated,
    ROUND(
        CASE 
            WHEN max_monthly_cl_last > 0 THEN max_monthly_cl_last
            WHEN max_monthly_cl_pre_last > 0 THEN max_monthly_cl_pre_last
            ELSE mc_monthly_cl
        END, 0
    ) AS mc_cl_migrated,
    ROUND(
        CASE 
            WHEN max_monthly_cl_last > 0 THEN max_monthly_cl_last * mc_cl_rate * 18
            WHEN max_monthly_cl_pre_last > 0 THEN max_monthly_cl_pre_last * mc_cl_rate * 18
            ELSE mc_monthly_cl * mc_cl_rate * 18
        END, 0
    ) AS mylo_rev18_cl
FROM 
    FinalResults;
'''

cl_migration_df = dataFetchingFunc(
    Query=query,
    SERVER="BI-DR-DB",
    DATABASE="master"
)

cl_migration_df = cl_migration_df.astype({'mc_cl_rate': 'float64'})
cl_migration_df = cl_migration_df.astype({'is_mylo_migrated': 'bool'})

cl_migration_df.head()

cl_migration_df.to_parquet(data_path+f'analytics/{analysis_date}_cl_migration_df.parquet')
